Update README.md
# pbtc-christain-ecommerce
This is a **front-end landing page clone project** I built from a Figma design of the PBTC Christian eCommerce site. It was created purely for **practice** and to showcase on my **portfolio** — this is not a live website or client work.   It includes basic interactivity such as a working mobile menu and a simple product card slider.